+++ packages/extension-polkagate/src/fullscreen/governance/tracks/util.ts
@@ -1,8 +1,9 @@
 // Copyright 2017-2024 @polkadot/extension-polkagate authors & contributors
 // SPDX-License-Identifier: Apache-2.0
-// @ts-nocheck
 
-import { BN, formatBalance } from '@polkadot/util';
+import type { BN } from '@polkadot/util';
+
+import { formatBalance } from '@polkadot/util';
 
 interface FormatOptions {
   decimals: number;
@@ -11,7 +12,7 @@ interface FormatOptions {
   withUnit: string;
 }
 
-export function formatSpendFactory(options: FormatOptions): (mul: number, value: BN) => string {
+export function formatSpendFactory (options: FormatOptions): (mul: number, value: BN) => string {
   return (mul: number, value: BN): string => {
     // We lose the decimals here... depending on chain config, this could be non-optimal
     // (A simple formatBalance(value.muln(mul), FMT_OPTS) formats to 4 decimals)
@@ -19,7 +20,7 @@ export function formatSpendFactory(options: FormatOptions): (mul: number, value:
   };
 }
 
-export function compareFellowshipRank(trackId: number): (rank: BN) => boolean {
+export function compareFellowshipRank (trackId: number): (rank: BN) => boolean {
   return (rank: BN): boolean =>
     rank.gten(trackId);
 }