+++ packages/extension-polkagate/src/util/types.ts
@@ -615,6 +615,8 @@ export interface PricesInCurrencies {
 
 export interface Price {
   price: number;
+  decimal: number;
+  token: string;
   priceChainName: string;
   priceDate: number;
 }